add sync functionality to apps.toml (#36)
- running install.sh will print a warning if there are installed apps
which are not present in apps.toml

## Summary by Sourcery

Add synchronization functionality to the install script to warn users if
there are installed apps not listed in apps.toml. Update terminology to
refer to 'apps and packages' instead of 'apps and software' for
consistency.

New Features:
- Add synchronization functionality to check for installed apps not
present in apps.toml.

Enhancements:
- Update terminology from 'apps and software' to 'apps and packages' for
consistency.