feat: Create bootstrap CI check (#52)
- create bootstrap.yaml GH action
- cleanup and linting

## Summary by Sourcery

Add a new GitHub Action workflow to automate the bootstrapping of
dotfiles on macOS and Ubuntu, and enhance the bootstrap script to
support branch-specific cloning. Update the CI configuration to include
a new bootstrap check and enable bashism checks in the linting process.

New Features:
- Introduce a new GitHub Action workflow 'bootstrap.yaml' to automate
the bootstrapping process for dotfiles on macOS and Ubuntu.

Enhancements:
- Enhance the bootstrap script to support cloning a specific branch of
the dotfiles repository.

CI:
- Add a new CI check with 'bootstrap.yaml' to ensure the bootstrapping
process works on both macOS and Ubuntu.
- Enable 'sh_checker_checkbashisms' in the linting workflow to check for
bashisms.